restructured the folder structure
