+++ frontend/utils/global.tsx
